Add tofloat(Point)
It is easy to go from a number to a Geometry.Point, but it is not easy
to go in the other direction (unless the type of point of point is
known). This PR adds a new function `tofloat` to X/Y/Z/Lat/Long point to
convert them into numbers.

This feature would be very useful in converting a Field of ZPoints into
a scalar field, which is what we need for the interpolation routines