Add interpolate(field, hcoords, zcoords)
Remapping a Field to a Cartesian grid is a laborious process. One needs
to define the target coordinates (which have to be `Geometry.Points`).
Then, they need to create a `Remapper`. Finally, one can use the
`Remapper` to interpolate.

This PR provides a way to skip one of these steps, the creation of the
`Remapper`, and offers a way to go from Field + coordinates to array.

This is particularly useful when working interactively because it
enables a workflow where one defines the target coordinates once, but
they can freely change the space and still obtain their target array
with `interpolate(field, hcoords, zcoords)`