Add `face/center_space`
As I started using the `CommonSpaces`, I realized that specifying the
staggering was a pain point. It required importing `Grids`, and I pretty
much always wanted to build CenterSpaces anyways. In addition to this, I
realized that going from center to faces and viceversa was not trivial.
This required carrying around both the center and the face space when
only one is really needed. In this commit, I reexport the staggering in
`CommonSpaces` and a function to go from center to face spaces and back.