Use TestEnv for downstream tests
Previously, there was a special case for ClimaTimeSteppers, where
the environment decsribed by test/Project.toml was used. This caused
errors because the Project.toml file was removed. So, the temporary test
environment generated wth Pkg.test() must be used. This does not work,
because it requires dev-ing ClimaCore from the ClimaTimeSteppers project,
which causes an Aqua test to fail.

ClimaUtilities was not passing because dev-ing ClimaCore
caused aqua tests to fail. A possible solution to this is
making a special case, similar to what was done for ClimaTimeSteppers.

I attempted to dev ClimaCore in the temporary test env generated
by Pkg.Test() by passing in `-e using Pkg; Pkg.dev(...)` as a julia
arg to Pkg.test, but that did not work.

Julia 1.11 adds a kwarg to Pkg.add, which specifies the field
inside the Project.toml file that a package should be added to.
If we only used 1.11, we could add the path to ClimaCore and specify
the target as extra or weakdeps, which would allow aqua tests to pass.

Instead, this commit makes the downstream test install TestEnv
into the base julia env, then activate the test env for the downstream
package, then dev ClimaCore, and then run the test suite.