Add support for InputOuput with PointSpaces
`write!` on Fields is now dispatched off the input
field's space. When it is a point space, the local
geometry data is written. When read_field is called,
and no grid is present, then the field is assumed
to be on a point space.

Add function comments

Simplify unit_point read/write test

Make suggested changes to unit_point and docstrings

clean up point space test