Add `read_type` to read field eltypes
When the reader recreates a field from a HDF5, it called
eval on a parsed input string. This replaces that with a function
that ensures there are no function calls inside the parsed string
before eval'ing it. This is based on old HDF5.jl code.

make suggested read_type changes

Revert field_eltype to value_type