Add support for HDF5 read and write with PointSpaces (#2162)
* Add support for InputOuput with PointSpaces

`write!` on Fields is now dispatched off the input
field's space. When it is a point space, the local
geometry data is written. When read_field is called,
and no grid is present, then the field is assumed
to be on a point space.

Add function comments

Simplify unit_point read/write test

Make suggested changes to unit_point and docstrings

clean up point space test

* Add `read_type` to read field eltypes

When the reader recreates a field from a HDF5, it called
eval on a parsed input string. This replaces that with a function
that ensures there are no function calls inside the parsed string
before eval'ing it. This is based on old HDF5.jl code.

make suggested read_type changes

Revert field_eltype to value_type

* Change value_type attribute to field_eltype

Change the value_type attribute, used by the HDF5 reader
and writer, to field_eltype. Backwards support is maintained
and tested.

Remove HDF5 import from test

import climacomms backend