Prevent mask from being incorrectly used
`Remapping.Remapper` performs spectral interpolation. When interpolating, every
point in a given element is used to calculate the target value. As a result, the
result is incorrect if one or more of the nodes in an element are masked.

This commit prevents users from accidentally obtaining subtly incorrect results
by calling the Remapper on a space that is not supported.

Note that if none of the nodes in an element are masked, the interpolation
routine would be fine. A more refined implementation of this check could take
this into consideration.